Add files via upload
Correction to splice to allow get_iterator and is_active to work
Corrections to get_iterator and is_active in situations where the iterator/pointer points to a block which has been removed then added back into the iterative sequence (via either unused_groups or an allocator), but now points to an element after end().
Minor optimization to splice.

+++ plf_hive_test_suite.cpp
@@ -1770,7 +1770,7 @@ int main()
 				hive1.erase(--(hive1.end()));
 				hive2.erase(--(hive2.end()));
 
-				const unsigned int total = hive1.size() + hive2.size();
+				const unsigned int total = static_cast<unsigned int>(hive1.size() + hive2.size());
 
 				hive1.splice(hive2);
